完成lab5测试1

+++ src/user/initcode.c
@@ -5,9 +5,9 @@ int main()
 {
     int L[5];
     char* s = "hello, world"; 
-    syscall(SYS_copyout, L);
-    syscall(SYS_copyin, L, 5);
-    syscall(SYS_copyinstr, s);
+    syscall(SYS_copyout, (long)L);
+    syscall(SYS_copyin, (long)L, 5);
+    syscall(SYS_copyinstr, (long)s);
     while(1);
     return 0;
 }
